refactor(logic): extract combat constants and engagement types (Refs #2288)
Move top-level combat resolution constants into combat_constants.dart and
EngagementResult / EngagementOutcome into combat_types.dart. combat_resolver.dart
imports them and re-exports for unchanged barrel/transitive consumers.

Combat resolver is now under the 700-line split threshold (remaining extractions
in issue step 3 can follow).
